tambah model

+++ database/migrations/2026_05_04_032952_create_wali_kelas_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->unsignedBigInteger('kelas_id');
             $table->unsignedBigInteger('user_id');
-            $table->int('nip');
+            $table->bigInteger('nip');
             $table->string('no_hp');
